+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## CHANGELOG FOR 1.2.x
 #### 1.2.0
+- feature  [#1637](https://github.com/ergonode/backend/pull/1637) Add posibility to import multimedia (rprzedzik) 
 - feature [#1645](https://github.com/ergonode/backend/pull/1645) Add attribute code validation (wiewiurdp)
 - bugfix [#1636](https://github.com/ergonode/backend/issues/1636) Multimedia gird is too slow  (wfajczyk)
 - bugfix [#1625](https://github.com/ergonode/backend/issues/1625) Allowed empty values for deactivated scheduler (piotrkreft)